\name{epmgpr-package}
\alias{epmgpr-package}
\alias{epmgpr}
\docType{package}
\title{
\packageTitle{epmgpr}
}
\description{
\packageDescription{epmgpr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{epmgpr}
\packageIndices{epmgpr}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{epmgpr}

Maintainer: \packageMaintainer{epmgpr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
